Clean up training_pane.
